Extract db-conn-setup middleware for prepare-context
