+++ lib/cartodb/controllers/analyses.js
@@ -1,9 +1,11 @@
 var PSQL = require('cartodb-psql');
 var cors = require('../middleware/cors');
 var userMiddleware = require('../middleware/user');
+const dbConnSetup = require('../middleware/db-conn-setup');
 
-function AnalysesController(prepareContext) {
+function AnalysesController(prepareContext, pgConnection) {
     this.prepareContext = prepareContext;
+    this.pgConnection = pgConnection;
 }
 
 module.exports = AnalysesController;
@@ -14,6 +16,7 @@ AnalysesController.prototype.register = function (app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createPGClient(),
         getDataFromQuery({ queryTemplate: catalogQueryTpl, key: 'catalog' }),
         getDataFromQuery({ queryTemplate: tablesQueryTpl, key: 'tables' }),
