+++ lib/cartodb/controllers/layergroup.js
@@ -2,6 +2,7 @@ const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
 const allowQueryParams = require('../middleware/allow-query-params');
 const vectorError = require('../middleware/vector-error');
+const dbConnSetup = require('../middleware/db-conn-setup');
 const DataviewBackend = require('../backends/dataview');
 const AnalysisStatusBackend = require('../backends/analysis-status');
 const MapStoreMapConfigProvider = require('../models/mapconfig/provider/map-store-provider');
@@ -55,6 +56,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getTile(this.tileBackend, 'map_tile'),
         setCacheControlHeader(),
@@ -74,6 +76,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getTile(this.tileBackend, 'map_tile'),
         setCacheControlHeader(),
@@ -94,6 +97,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getTile(this.tileBackend, 'maplayer_tile'),
         setCacheControlHeader(),
@@ -113,6 +117,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getFeatureAttributes(this.attributesBackend),
         setCacheControlHeader(),
@@ -131,6 +136,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(['layer']),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi, forcedFormat),
         getPreviewImageByCenter(this.previewBackend),
         setCacheControlHeader(),
@@ -147,6 +153,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(['layer']),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi, forcedFormat),
         getPreviewImageByBoundingBox(this.previewBackend),
         setCacheControlHeader(),
@@ -181,6 +188,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getDataview(this.dataviewBackend),
         setCacheControlHeader(),
@@ -197,6 +205,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         getDataview(this.dataviewBackend),
         setCacheControlHeader(),
@@ -213,6 +222,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         dataviewSearch(this.dataviewBackend),
         setCacheControlHeader(),
@@ -229,6 +239,7 @@ LayergroupController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(allowedDataviewQueryParams),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         createMapStoreMapConfigProvider(this.mapStore, this.userLimitsApi),
         dataviewSearch(this.dataviewBackend),
         setCacheControlHeader(),
@@ -244,6 +255,7 @@ LayergroupController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         analysisNodeStatus(this.analysisStatusBackend),
         sendResponse()
     );
