+++ lib/cartodb/controllers/map.js
@@ -7,6 +7,7 @@ const ResourceLocator = require('../models/resource-locator');
 const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
 const allowQueryParams = require('../middleware/allow-query-params');
+const dbConnSetup = require('../middleware/db-conn-setup');
 const NamedMapsCacheEntry = require('../cache/model/named_maps_entry');
 const NamedMapMapConfigProvider = require('../models/mapconfig/provider/named-map-provider');
 const CreateLayergroupMapConfigProvider = require('../models/mapconfig/provider/create-layergroup-provider');
@@ -69,6 +70,7 @@ MapController.prototype.composeCreateMapMiddleware = function (useTemplate = fal
         userMiddleware(),
         allowQueryParams(['aggregation']),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         initProfiler(isTemplateInstantiation),
         checkJsonContentType(),
         this.getCreateMapMiddlewares(useTemplate),
