+++ lib/cartodb/controllers/named_maps.js
@@ -1,6 +1,7 @@
 const NamedMapsCacheEntry = require('../cache/model/named_maps_entry');
 const cors = require('../middleware/cors');
 const userMiddleware = require('../middleware/user');
+const dbConnSetup = require('../middleware/db-conn-setup');
 const allowQueryParams = require('../middleware/allow-query-params');
 const vectorError = require('../middleware/vector-error');
 
@@ -28,7 +29,7 @@ function getRequestParams(locals) {
 }
 
 function NamedMapsController(prepareContext, namedMapProviderCache, tileBackend, previewBackend,
-                             surrogateKeysCache, tablesExtentApi, metadataBackend) {
+                             surrogateKeysCache, tablesExtentApi, metadataBackend, pgConnection) {
     this.namedMapProviderCache = namedMapProviderCache;
     this.tileBackend = tileBackend;
     this.previewBackend = previewBackend;
@@ -36,6 +37,7 @@ function NamedMapsController(prepareContext, namedMapProviderCache, tileBackend,
     this.tablesExtentApi = tablesExtentApi;
     this.metadataBackend = metadataBackend;
     this.prepareContext = prepareContext;
+    this.pgConnection = pgConnection;
 }
 
 module.exports = NamedMapsController;
@@ -48,6 +50,7 @@ NamedMapsController.prototype.register = function(app) {
         cors(),
         userMiddleware(),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         getNamedMapProvider({
             namedMapProviderCache: this.namedMapProviderCache,
             label: 'NAMED_MAP_TILE'
@@ -72,6 +75,7 @@ NamedMapsController.prototype.register = function(app) {
         userMiddleware(),
         allowQueryParams(['layer', 'zoom', 'lon', 'lat', 'bbox']),
         this.prepareContext,
+        dbConnSetup(this.pgConnection),
         getNamedMapProvider({
             namedMapProviderCache: this.namedMapProviderCache,
             label: 'STATIC_VIZ_MAP', forcedFormat: 'png'
