+++ lib/cartodb/middleware/context/index.js
@@ -3,15 +3,13 @@ const cleanUpQueryParams = require('./clean-up-query-params');
 const layergroupToken = require('./layergroup-token');
 const credentials = require('./credentials');
 const authorize = require('./authorize');
-const dbConnSetup = require('./db-conn-setup');
 
-module.exports = function prepareContextMiddleware(authApi, pgConnection) {
+module.exports = function prepareContextMiddleware(authApi) {
     return [
         locals(),
         cleanUpQueryParams(),
         layergroupToken(),
         credentials(),
-        authorize(authApi),
-        dbConnSetup(pgConnection)
+        authorize(authApi)
     ];
 };
