+++ lib/cartodb/server.js
@@ -217,7 +217,7 @@ module.exports = function(serverOptions) {
 
     const prepareContext = typeof serverOptions.req2params === 'function' ?
         serverOptions.req2params :
-        prepareContextMiddleware(authApi, pgConnection);
+        prepareContextMiddleware(authApi);
 
     /*******************************************************************************************************************
      * Routing
@@ -256,12 +256,13 @@ module.exports = function(serverOptions) {
         previewBackend,
         surrogateKeysCache,
         tablesExtentApi,
-        metadataBackend
+        metadataBackend,
+        pgConnection
     ).register(app);
 
     new controller.NamedMapsAdmin(authApi, templateMaps).register(app);
 
-    new controller.Analyses(prepareContext).register(app);
+    new controller.Analyses(prepareContext, pgConnection).register(app);
 
     new controller.ServerInfo(versions).register(app);
